+++ packages/javascript/lib/object/index.ts
@@ -1,6 +1,7 @@
 export { objDeleteKeyByValue } from './objDeleteKeyByValue.js';
 export { objFindItemRecursiveByKey } from './objFindItemRecursiveByKey.js';
 export { objMergeNewKey } from './objMergeNewKey.js';
+export { objPickBy } from './objPickBy.js';
 export { objTo1d } from './objTo1d.js';
 export { objToArray } from './objToArray.js';
 export { objToPrettyStr } from './objToPrettyStr.js';